feat(search): 新增商品搜索功能实现
- 创建了Attr DTO类用于属性数据传输
- 配置了网关路由规则支持各微服务访问
- 添加了产品服务的Redis缓存配置
- 实现了Elasticsearch搜索常量定义
- 创建了首页模板文件提供商品展示界面
- 定义了搜索参数和结果的数据结构
- 实现了完整的商品搜索服务逻辑包括查询构建、聚合分析和结果封装

+++ gulimall-Parent/services/gulimall-product/src/main/java/com/xjz/gulimall/product/dto/Attr.java
@@ -13,7 +13,7 @@ import lombok.Data;
  */
 @Data
 public class Attr {
-    public int attrId;
+    public Long attrId;
     public String attrName;
     public String attrValue;
 }
